move App class into app.js
